Meshの生成失敗

+++ DirectX_Render_Study/DirectX12_BoxDraw_Refactoring/ModelData.h
@@ -19,7 +19,7 @@ struct MeshData
     std::string name;
 
     std::vector<MeshVertex> vertices;
-    std::vector<uint32_t> indices;
+    std::vector<uint16_t> indices;
 
     uint32_t materialIndex;
 };
